third_party/WebKit/LayoutTests/fast/dom/Element/attr-param-typechecking-expected.txt

This test checks whether passing wrong types to setAttributeNode causes a crash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.setAttributeNode("style"); threw exception TypeMismatchError: Failed to execute 'setAttributeNode' on 'Element': The 1st argument provided is either null, or an invalid Attr object..
PASS element.setAttributeNode(null); threw exception TypeMismatchError: Failed to execute 'setAttributeNode' on 'Element': The 1st argument provided is either null, or an invalid Attr object..
PASS element.setAttributeNode(undefined); threw exception TypeMismatchError: Failed to execute 'setAttributeNode' on 'Element': The 1st argument provided is either null, or an invalid Attr object..
PASS element.setAttributeNode(new Object); threw exception TypeMismatchError: Failed to execute 'setAttributeNode' on 'Element': The 1st argument provided is either null, or an invalid Attr object..
PASS element.removeAttributeNode("style"); threw exception TypeMismatchError: Failed to execute 'removeAttributeNode' on 'Element': The 1st argument provided is either null, or an invalid Attr object..
PASS element.removeAttributeNode(null); threw exception TypeMismatchError: Failed to execute 'removeAttributeNode' on 'Element': The 1st argument provided is either null, or an invalid Attr object..
PASS element.removeAttributeNode(undefined); threw exception TypeMismatchError: Failed to execute 'removeAttributeNode' on 'Element': The 1st argument provided is either null, or an invalid Attr object..
PASS element.removeAttributeNode(new Object); threw exception TypeMismatchError: Failed to execute 'removeAttributeNode' on 'Element': The 1st argument provided is either null, or an invalid Attr object..
PASS successfullyParsed is true

TEST COMPLETE
